requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.5
djangorestframework==3.10
djangorestframework-simplejwt==5.2.2
django-cors-headers
pyjwt
idna==3.4
#psycopg2==2.9.5
gunicorn==20.1.0
python-nginx==1.5.7
psycopg2-binary==2.9.5
PyQRCode==1.2.1
pytz==2022.7
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
tqdm==4.64.1
urllib3==1.26.13